+++ src/components/contetnt/Shirts/Shirts.jsx
@@ -1,18 +1,23 @@
-import { Col, Container, Row } from "react-bootstrap"
-import { CardShirts } from "../Cards/CardShirts"
+import { Container, Row } from "react-bootstrap"
+import { Cards } from "../Cards/Card"
+import data from "../Cards/data.json"
+
 
 
 export const Shirts = () => {
-    let num = [0, 1, 2, 3, 4, 5, 6, 7]
-    let idList = num.map((id) => <Col><CardShirts id={id} /></Col>)
+
+const { shirts } = data
+
     return (
         <>
-    <h2>Shirts</h2>
-    <Container>
-        <Row xs={1} md={2} lg={3} xl={4} className="g-3">
-            {idList}
-        </Row>
-    </Container>
-    </>
+            <h2>Shirts</h2>
+            <Container>
+                <Row xs={1} md={2} lg={3} xl={4} className="g-3">
+                {shirts.map((card) => (
+                    <Cards card={card} />
+                ))}
+                </Row>
+            </Container>
+        </>
     )
 }
