+++ src/components/contetnt/Tickets/Tickets.jsx
@@ -1,18 +1,24 @@
-import { Col, Container, Row } from "react-bootstrap"
-import { CardTicket } from "../Cards/CardTicket"
+import { Container, Row } from "react-bootstrap"
+import { Cards } from "../Cards/Card"
+
+
+import data from "../Cards/data.json"
 
 
 export const Tickets = () => {
-    let num = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11]
-    let idList = num.map((id) => <Col><CardTicket id={id} /></Col>)
+     const { eTicket } = data
+
     return (
         <>
-    <h2>Tickets</h2>
-    <Container>
-        <Row xs={1} md={2} lg={3} xl={4} className="g-3">
-            {idList}
-        </Row>
-    </Container>
+            <h2>Tickets</h2>
+            <Container>
+                <Row xs={1} md={2} lg={3} xl={4} className="g-3">
+                    {eTicket.map((card) => (
+                        <Cards card={card} />
+                    ))}
+                </Row>
+            </Container>
     </>
+
     )
 }
